Consolidated everything except prob doublets into doubletdetectionchanged some comments
