--- utils.py
@@ -1,58 +1,0 @@
-"""Data handling utilities."""
-import numpy as np
-import pandas as pd
-
-
-def load_data(FNAME, normalize=False, read_rows=None):
-    """Load a csv table from the filesystem.
-
-    Args:
-        FNAME (string): Pathname of file to load.
-        normalize (bool, optional): Runs normalize_counts on table.
-        read_rows (None or int, optional): If specified, load only first
-            read_rows of FNAME.
-
-    Returns:
-        ndarray: Loaded table.
-    """
-    counts = pd.read_csv(FNAME, index_col=0, nrows=read_rows).as_matrix()
-
-    if normalize:
-        counts = normalize_counts(counts)
-
-    return counts
-
-
-def normalize_counts(raw_counts, doStandardize=False):
-    """Normalize count array using method in 10x pipeline.
-
-    From http://www.nature.com/articles/ncomms14049.
-
-    Args:
-        raw_counts (ndarray): count data
-        doStandardize (bool, optional): Standardizes each gene column.
-
-    Returns:
-        ndarray: Normalized data.
-    """
-    # Sum across cells
-    cell_sums = np.sum(raw_counts, axis=1)
-
-    # Mutiply by median and divide each cell by cell sum
-    median = np.median(cell_sums)
-    raw_counts = raw_counts * median / cell_sums[:, np.newaxis]
-
-    raw_counts = np.log(raw_counts + 0.1)
-
-    if doStandardize:
-        # Normalize to have genes with mean 0 and std 1
-        std = np.std(raw_counts, axis=0)[np.newaxis, :]
-
-        # Fix potential divide by zero
-        std[np.where(std == 0)[0]] = 1
-
-        normed = (raw_counts - np.mean(raw_counts, axis=0)) / std
-    else:
-        normed = raw_counts
-
-    return normed
